refactor: categories module folder
