+++ packages/shared/src/categories/category.schema.ts
@@ -6,11 +6,17 @@ export const createCategorySchema = z.object({
   description: z.string().trim().max(200).optional(),
 });
 
+export const categorySchema = z.object({
+  id: z.string(),
+  name: z.string(),
+  slug: z.string(),
+  description: z.string().optional(),
+  createdAt: z.string(),
+  updatedAt: z.string(),
+});
+
 export const categorySummarySchema = z.object({
   id: z.string(),
   name: z.string(),
   slug: z.string(),
 });
-
-export type CreateCategoryBody = z.infer<typeof createCategorySchema>;
-export type CategorySummary = z.infer<typeof categorySummarySchema>;
