+++ packages/shared/src/index.ts
@@ -1,10 +1,16 @@
+// biome-ignore-all assist/source/organizeImports: reason for ignoring
+
 export * from "./auth/auth.schema.js";
-export * from "./auth/auth.types.js";
-export * from "./category.schema.js";
+export type * from "./auth/auth.types.js";
+
+export * from "./categories/category.schema.js";
+export type * from "./categories/category.types.js";
+
+export * from "./users/user.schema.js";
+export type * from "./users/user.types.js";
+
 export * from "./comment.schema.js";
 export * from "./pagination.js";
 export * from "./recipe.schema.js";
 export * from "./types.js";
-export * from "./users/user.schema.js";
-export * from "./users/user.types.js";
 export * from "./utils.js";
