+++ packages/shared/src/recipe.schema.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { categorySummarySchema } from "./category.schema.js";
+import { categorySummarySchema } from "./categories/category.schema.js";
 import { userSummarySchema } from "./users/user.schema.js";
 
 export const minutesSchema = z.number().int().min(1).brand<"Minutes">();
